main 왼쪽부분 컴포넌트화 및 피드mockdata 받아 랜더링 성공
