+++ src/pages/taewonKim/Main/FeedRight/FeedRight.scss
@@ -1,4 +1,4 @@
-.feed-sub {
+.feed-right {
   width: 220px;
 
   .sub-me {
