+++ src/pages/taewonKim/Main/FeedRight/Story/Story.scss
