+++ src/pages/taewonKim/Main/Main.scss
@@ -84,187 +84,14 @@
 
 .main-container {
   display: flex;
-  justify-content: space-between;
-  width: 720px;
-  margin: 50px auto 0 auto;
+  justify-content: center;
+  padding-top: 50px;
+  gap: 40px;
 
-  .feed {
-    height: 100%;
-    border: 1px solid #e5e5e5;
-    background-color: #fff;
-
-    .feed-top {
-      display: flex;
-      justify-content: space-between;
-      height: 50px;
-      padding: 0 10px;
-
-      .user-info {
-        display: flex;
-        align-items: center;
-
-        .photo {
-          width: 24px;
-          height: 24px;
-          margin-right: 10px;
-          border-radius: 50%;
-        }
-
-        .nick-name {
-          font-size: 12px;
-        }
-      }
-
-      .more-btn {
-        display: flex;
-        align-items: center;
-        border: 0;
-        background-color: rgba(0, 0, 0, 0);
-
-        .more {
-          width: 21px;
-        }
-      }
-    }
-
-    .feed-main {
-      .feed-img {
-        width: 468px;
-        height: 468px;
-      }
-    }
-
-    .feed-bottom {
-      .feed-info-box {
-        display: flex;
-        justify-content: space-between;
-        height: 30px;
-        margin-bottom: 5px;
-        padding: 0 10px;
-
-        .feed-info-left {
-          display: flex;
-          justify-content: space-between;
-          width: 80px;
-
-          .feed-info-btn {
-            border: 0px;
-            background-color: #fff;
-
-            .feed-like-icon,
-            .feed-comment-icon,
-            .feed-share-icon {
-              width: 21px;
-            }
-          }
-        }
-
-        .feed-info-right {
-          .feed-info-btn {
-            border: 0px;
-            background-color: #fff;
-
-            .feed-bookmark-icon {
-              width: 28px;
-            }
-          }
-        }
-      }
-
-      .like {
-        display: flex;
-        align-items: center;
-        padding: 0 10px;
-
-        .photo {
-          width: 24px;
-          height: 24px;
-          margin-right: 10px;
-          border-radius: 50%;
-        }
-
-        .nick-name {
-          font-size: 12px;
-        }
-      }
-
-      .feed-text-box {
-        margin: 10px 0;
-        padding: 0 10px;
-
-        .feed-text {
-          font-size: 12px;
-        }
-
-        .upload-time {
-          color: #9f9f9f;
-          font-size: 8px;
-        }
-      }
-
-      .post-comment-box {
-        padding-left: 10px;
-
-        .comment-list {
-          display: flex;
-          position: relative;
-          align-items: center;
-          height: 25px;
-
-          .comment-user {
-            font-size: 12px;
-            font-weight: bold;
-          }
-
-          .comment-text {
-            margin-left: 10px;
-            font-size: 12px;
-          }
-
-          .comment-like-btn {
-            position: absolute;
-            right: 15px;
-            border: 0;
-            background-color: rgba(0, 0, 0, 0);
-
-            .comment-like-icon {
-              width: 15px;
-            }
-          }
-        }
-      }
-
-      .comment-box {
-        position: relative;
-        border-top: 1px solid #e5e5e5;
-
-        .comment {
-          width: 100%;
-          height: 35px;
-          padding: 6px 50px 6px 10px;
-          border: 0px;
-        }
-
-        .post-btn {
-          position: absolute;
-          top: 0;
-          right: 10px;
-          width: max-content;
-          height: 100%;
-          border: 0px;
-          background-color: rgba(255, 255, 255, 0);
-          color: blue;
-        }
-
-        .post-btn-default {
-          opacity: 0.3;
-        }
-
-        .post-btn-active {
-          opacity: 1;
-          cursor: pointer;
-        }
-      }
-    }
+  .article-container {
+    display: flex;
+    flex-direction: column;
+    width: max-content;
+    gap: 20px;
   }
 }
